+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* Removed Tessa gallery button
+
 ## 1.3.4 (2022-11-29)
 * Fixed dependencies for Community Edition which prevented the installation